+++ mautrix/bridge/matrix.py
@@ -501,18 +501,20 @@ class BaseMatrixHandler:
     async def handle_receipt(self, evt: ReceiptEvent) -> None:
         for event_id, receipts in evt.content.items():
             for user_id, data in receipts[ReceiptType.READ].items():
-                if (
-                    data.get(DOUBLE_PUPPET_SOURCE_KEY) == self.az.bridge_name
-                    and await self.bridge.get_double_puppet(user_id) is not None
-                ):
-                    continue
-
                 user = await self.bridge.get_user(user_id, create=False)
                 if not user or not await user.is_logged_in():
                     continue
 
                 portal = await self.bridge.get_portal(evt.room_id)
                 if not portal:
+                    continue
+
+                await portal.schedule_disappearing()
+
+                if (
+                    data.get(DOUBLE_PUPPET_SOURCE_KEY) == self.az.bridge_name
+                    and await self.bridge.get_double_puppet(user_id) is not None
+                ):
                     continue
 
                 await self.handle_read_receipt(user, portal, event_id, data)